+++ src/state.rs
@@ -14,7 +14,7 @@ pub const STATE: Item<State> = Item::new("state");
 #[cw_serde]
 pub struct Config {
     /// Number of tokens that has been registered.
-    pub registered_tokens: u128,
+    pub registered_tokens: u64,
     /// Base URI for metadata. It is immutable after the creation.
     pub metadata_uri: String,
     /// When the minter value is changed to Empty, no token can be minted
